requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymysql==1.1.0
pydantic==2.7.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
Pillow>=10.2.0
